% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans.R
\docType{class}
\name{Kmeans}
\alias{Kmeans}
\title{K-means for variable clustering (numeric variables only)}
\description{
K-means for variable clustering (numeric variables only)

K-means for variable clustering (numeric variables only)
}
\details{
R6 class implementing a k-means-like algorithm for clustering
\emph{variables} (columns) when all active variables are numeric.
}
\section{Super class}{
\code{\link[mmrClustVar:ClusterBase]{mmrClustVar::ClusterBase}} -> \code{Kmeans}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Kmeans-new}{\code{Kmeans$new()}}
\item \href{#method-Kmeans-interpret_clusters}{\code{Kmeans$interpret_clusters()}}
\item \href{#method-Kmeans-clone}{\code{Kmeans$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="fit"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-fit'><code>mmrClustVar::ClusterBase$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_K"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_K'><code>mmrClustVar::ClusterBase$get_K()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_X_descr"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_X_descr'><code>mmrClustVar::ClusterBase$get_X_descr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_centers"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_centers'><code>mmrClustVar::ClusterBase$get_centers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_clusters"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_clusters'><code>mmrClustVar::ClusterBase$get_clusters()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_convergence"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_convergence'><code>mmrClustVar::ClusterBase$get_convergence()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_inertia"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_inertia'><code>mmrClustVar::ClusterBase$get_inertia()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_method"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_method'><code>mmrClustVar::ClusterBase$get_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="plot"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-plot'><code>mmrClustVar::ClusterBase$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="predict"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-predict'><code>mmrClustVar::ClusterBase$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="print"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-print'><code>mmrClustVar::ClusterBase$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="summary"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-summary'><code>mmrClustVar::ClusterBase$summary()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kmeans-new"></a>}}
\if{latex}{\out{\hypertarget{method-Kmeans-new}{}}}
\subsection{Method \code{new()}}{
Create a new Kmeans instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmeans$new(K, scale = TRUE, random_state = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{K}}{Number of cluster}

\item{\code{scale}}{A boolean defining whether to scale the data or not}

\item{\code{random_state}}{The random seed to use}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kmeans-interpret_clusters"></a>}}
\if{latex}{\out{\hypertarget{method-Kmeans-interpret_clusters}{}}}
\subsection{Method \code{interpret_clusters()}}{
Textual interpretation of the k-means variable clustering solution.

For each cluster, this method reports:
\itemize{
\item the number of variables (all numeric),
\item a few most representative variables (highest r^2 to the local PC),
\item simple membership statistics based on r^2.
}

The "distance" is defined as 1 - r^2(x_j, PC_k),
and the "adhesion" as r^2(x_j, PC_k).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmeans$interpret_clusters(style = c("compact", "detailed"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Character, either "compact" or "detailed".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kmeans-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Kmeans-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmeans$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
