% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmedoids.R
\docType{class}
\name{Kmedoids}
\alias{Kmedoids}
\title{K-medoids for variable clustering (mixed variables)}
\description{
K-medoids for variable clustering (mixed variables)

K-medoids for variable clustering (mixed variables)
}
\details{
R6 class implementing a k-medoids-like algorithm for clustering
\emph{variables} (columns) based on a dissimilarity matrix:
\itemize{
\item numeric vs numeric: 1 - r^2(x_j, x_l)
\item categorical vs categorical: simple matching dissimilarity
\item mixed (numeric vs categorical): fixed dissimilarity = 1
}
}
\section{Super class}{
\code{\link[mmrClustVar:ClusterBase]{mmrClustVar::ClusterBase}} -> \code{Kmedoids}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Kmedoids-new}{\code{Kmedoids$new()}}
\item \href{#method-Kmedoids-interpret_clusters}{\code{Kmedoids$interpret_clusters()}}
\item \href{#method-Kmedoids-clone}{\code{Kmedoids$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="fit"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-fit'><code>mmrClustVar::ClusterBase$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_K"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_K'><code>mmrClustVar::ClusterBase$get_K()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_X_descr"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_X_descr'><code>mmrClustVar::ClusterBase$get_X_descr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_centers"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_centers'><code>mmrClustVar::ClusterBase$get_centers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_clusters"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_clusters'><code>mmrClustVar::ClusterBase$get_clusters()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_convergence"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_convergence'><code>mmrClustVar::ClusterBase$get_convergence()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_inertia"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_inertia'><code>mmrClustVar::ClusterBase$get_inertia()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_method"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_method'><code>mmrClustVar::ClusterBase$get_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="plot"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-plot'><code>mmrClustVar::ClusterBase$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="predict"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-predict'><code>mmrClustVar::ClusterBase$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="print"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-print'><code>mmrClustVar::ClusterBase$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="summary"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-summary'><code>mmrClustVar::ClusterBase$summary()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kmedoids-new"></a>}}
\if{latex}{\out{\hypertarget{method-Kmedoids-new}{}}}
\subsection{Method \code{new()}}{
Create a new Kmedoids instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmedoids$new(K, lambda = 1, random_state = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{K}}{Number of cluster}

\item{\code{lambda}}{Weight for qualitative data}

\item{\code{random_state}}{The random seed to use}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kmedoids-interpret_clusters"></a>}}
\if{latex}{\out{\hypertarget{method-Kmedoids-interpret_clusters}{}}}
\subsection{Method \code{interpret_clusters()}}{
Textual interpretation of the k-medoids variable clustering solution.

For each cluster, this method reports:
\itemize{
\item the number of variables,
\item a few most central variables (closest to the medoid),
\item simple membership statistics based on distance to the medoid.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmedoids$interpret_clusters(style = c("compact", "detailed"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Character, either "compact" or "detailed".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kmedoids-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Kmedoids-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmedoids$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
