% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmodes.R
\docType{class}
\name{Kmodes}
\alias{Kmodes}
\title{K-modes for variable clustering (categorical variables only)}
\description{
K-modes for variable clustering (categorical variables only)

K-modes for variable clustering (categorical variables only)
}
\details{
R6 class implementing a k-modes-like algorithm for clustering
\emph{variables} (columns) when all active variables are categorical.
}
\section{Super class}{
\code{\link[mmrClustVar:ClusterBase]{mmrClustVar::ClusterBase}} -> \code{Kmodes}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Kmodes-new}{\code{Kmodes$new()}}
\item \href{#method-Kmodes-interpret_clusters}{\code{Kmodes$interpret_clusters()}}
\item \href{#method-Kmodes-clone}{\code{Kmodes$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="fit"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-fit'><code>mmrClustVar::ClusterBase$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_K"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_K'><code>mmrClustVar::ClusterBase$get_K()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_X_descr"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_X_descr'><code>mmrClustVar::ClusterBase$get_X_descr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_centers"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_centers'><code>mmrClustVar::ClusterBase$get_centers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_clusters"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_clusters'><code>mmrClustVar::ClusterBase$get_clusters()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_convergence"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_convergence'><code>mmrClustVar::ClusterBase$get_convergence()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_inertia"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_inertia'><code>mmrClustVar::ClusterBase$get_inertia()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_method"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_method'><code>mmrClustVar::ClusterBase$get_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="plot"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-plot'><code>mmrClustVar::ClusterBase$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="predict"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-predict'><code>mmrClustVar::ClusterBase$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="print"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-print'><code>mmrClustVar::ClusterBase$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="summary"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-summary'><code>mmrClustVar::ClusterBase$summary()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kmodes-new"></a>}}
\if{latex}{\out{\hypertarget{method-Kmodes-new}{}}}
\subsection{Method \code{new()}}{
Create a new Kmodes instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmodes$new(K, random_state = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{K}}{Number of cluster}

\item{\code{random_state}}{The random seed to use}

\item{\code{scale}}{A boolean defining whether to scale the data or not}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kmodes-interpret_clusters"></a>}}
\if{latex}{\out{\hypertarget{method-Kmodes-interpret_clusters}{}}}
\subsection{Method \code{interpret_clusters()}}{
Textual interpretation of the k-modes variable clustering solution.

For each cluster, this method reports:
\itemize{
\item the number of variables (categorical),
\item a few most representative variables (highest agreement with the
cluster modal profile),
\item simple membership statistics based on this agreement.
}

The "distance" is defined as 1 - proportion of matches with the
cluster-level modal profile, and "adhesion" as that proportion.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmodes$interpret_clusters(style = c("compact", "detailed"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Character, either "compact" or "detailed".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kmodes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Kmodes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmodes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
