% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kprototypes.R
\docType{class}
\name{Kprototypes}
\alias{Kprototypes}
\title{K-prototypes for variable clustering (mixed variables)}
\description{
K-prototypes for variable clustering (mixed variables)

K-prototypes for variable clustering (mixed variables)
}
\details{
R6 class implementing a k-prototypes-like algorithm for clustering
\emph{variables} (columns) when active variables are mixed (numeric +
categorical).
}
\section{Super class}{
\code{\link[mmrClustVar:ClusterBase]{mmrClustVar::ClusterBase}} -> \code{Kprototypes}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Kprototypes-new}{\code{Kprototypes$new()}}
\item \href{#method-Kprototypes-interpret_clusters}{\code{Kprototypes$interpret_clusters()}}
\item \href{#method-Kprototypes-clone}{\code{Kprototypes$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="fit"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-fit'><code>mmrClustVar::ClusterBase$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_K"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_K'><code>mmrClustVar::ClusterBase$get_K()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_X_descr"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_X_descr'><code>mmrClustVar::ClusterBase$get_X_descr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_centers"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_centers'><code>mmrClustVar::ClusterBase$get_centers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_clusters"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_clusters'><code>mmrClustVar::ClusterBase$get_clusters()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_convergence"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_convergence'><code>mmrClustVar::ClusterBase$get_convergence()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_inertia"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_inertia'><code>mmrClustVar::ClusterBase$get_inertia()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="get_method"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-get_method'><code>mmrClustVar::ClusterBase$get_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="plot"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-plot'><code>mmrClustVar::ClusterBase$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="predict"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-predict'><code>mmrClustVar::ClusterBase$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="print"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-print'><code>mmrClustVar::ClusterBase$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClustVar" data-topic="ClusterBase" data-id="summary"><a href='../../mmrClustVar/html/ClusterBase.html#method-ClusterBase-summary'><code>mmrClustVar::ClusterBase$summary()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kprototypes-new"></a>}}
\if{latex}{\out{\hypertarget{method-Kprototypes-new}{}}}
\subsection{Method \code{new()}}{
Create a new Kprototypes instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kprototypes$new(K, scale = TRUE, lambda = 1, random_state = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{K}}{Number of cluster}

\item{\code{scale}}{A boolean defining whether to scale the data or not}

\item{\code{lambda}}{Weight for qualitative data}

\item{\code{random_state}}{The random seed to use}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kprototypes-interpret_clusters"></a>}}
\if{latex}{\out{\hypertarget{method-Kprototypes-interpret_clusters}{}}}
\subsection{Method \code{interpret_clusters()}}{
Textual interpretation of the k-prototypes variable clustering solution.

For each cluster, this method reports:
\itemize{
\item the number of numeric and categorical variables,
\item a few most representative variables (highest adhesion to the
numeric and/or categorical prototype),
\item simple membership statistics.
}

For numeric variables, adhesion is based on r^2 to a local PC (as in k-means).
For categorical variables, adhesion is based on agreement with a cluster-level
modal profile (as in k-modes).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kprototypes$interpret_clusters(style = c("compact", "detailed"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Character, either "compact" or "detailed".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kprototypes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Kprototypes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kprototypes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
