% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_mmrClustVar_app}
\alias{run_mmrClustVar_app}
\title{Launch the mmrClustVar Shiny application}
\usage{
run_mmrClustVar_app()
}
\value{
This function is called for its side-effect (launching the app).
}
\description{
This function launches the Shiny app included in the package,
located under \code{inst/shiny/mmrClustVar_app}.
}
\examples{
\dontrun{
run_mmrClustVar_app()
}
}
