% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{Kmeans}
\alias{Kmeans}
\title{K-means}
\description{
Implementation of the K-means variable clustering algorithm.
}
\examples{
kmeans_model <- Kmeans$new(n_cluster=5)
result <- kmeans_model$fit(mtcars)
print(result$clusters)
}
\section{Super class}{
\code{mmrClutVar::Clustering} -> \code{K-means}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-K-means-new}{\code{Kmeans$new()}}
\item \href{#method-K-means-fit}{\code{Kmeans$fit()}}
\item \href{#method-K-means-visualize}{\code{Kmeans$visualize()}}
\item \href{#method-K-means-get.clusters}{\code{Kmeans$get.clusters()}}
\item \href{#method-K-means-get.center}{\code{Kmeans$get.center()}}
\item \href{#method-K-means-clone}{\code{Kmeans$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mmrClutVar" data-topic="Clustering" data-id="get.max_iter"><a href='../../mmrClutVar/html/Clustering.html#method-Clustering-get.max_iter'><code>mmrClutVar::Clustering$get.max_iter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClutVar" data-topic="Clustering" data-id="get.n_cluster"><a href='../../mmrClutVar/html/Clustering.html#method-Clustering-get.n_cluster'><code>mmrClutVar::Clustering$get.n_cluster()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClutVar" data-topic="Clustering" data-id="predict"><a href='../../mmrClutVar/html/Clustering.html#method-Clustering-predict'><code>mmrClutVar::Clustering$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mmrClutVar" data-topic="Clustering" data-id="set.n_cluster"><a href='../../mmrClutVar/html/Clustering.html#method-Clustering-set.n_cluster'><code>mmrClutVar::Clustering$set.n_cluster()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-K-means-new"></a>}}
\if{latex}{\out{\hypertarget{method-K-means-new}{}}}
\subsection{Method \code{new()}}{
Create a new K-means object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmeans$new(
  n_cluster = NULL,
  center = TRUE,
  scale = TRUE,
  max_iter = 300,
  random_seed = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_cluster}}{An optional integer specifying the number of clusters. If \code{NULL}, the model will determine the optimal number of clusters.}

\item{\code{center}}{A logical indicating whether to center the data (default is \code{TRUE}).}

\item{\code{scale}}{A logical indicating whether to scale the data (default is \code{TRUE}).}

\item{\code{max_iter}}{An integer specifying the maximum number of iterations for the K-means algorithm (default is 300).}

\item{\code{random_seed}}{An optional integer to set the random seed for reproducibility.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-K-means-fit"></a>}}
\if{latex}{\out{\hypertarget{method-K-means-fit}{}}}
\subsection{Method \code{fit()}}{
Fits the K-means model to data, automatically finding the best K if not provided
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmeans$fit(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A data.frame or matrix on which to perform clustering}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing the following components:
\itemize{
\item \code{clusters}: An integer vector indicating the cluster to which each point is allocated.
\item \code{centers}: A matrix with the coordinates of cluster centers.
\item \code{W}: The within-cluster sum of squares for the fitted clusters.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-K-means-visualize"></a>}}
\if{latex}{\out{\hypertarget{method-K-means-visualize}{}}}
\subsection{Method \code{visualize()}}{
Project clusters on factorial plane
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmeans$visualize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-K-means-get.clusters"></a>}}
\if{latex}{\out{\hypertarget{method-K-means-get.clusters}{}}}
\subsection{Method \code{get.clusters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmeans$get.clusters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-K-means-get.center"></a>}}
\if{latex}{\out{\hypertarget{method-K-means-get.center}{}}}
\subsection{Method \code{get.center()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmeans$get.center()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-K-means-clone"></a>}}
\if{latex}{\out{\hypertarget{method-K-means-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kmeans$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
