% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_mmrClustVar_app}
\alias{run_mmrClustVar_app}
\title{Launch the demo Shiny application for mmrClustVar}
\usage{
run_mmrClustVar_app()
}
\value{
Nothing. Opens a Shiny app.
}
\description{
Launches a Shiny application demonstrating the use of the \code{mmrClustVar} R6 class for clustering variables.
}
\examples{
\dontrun{
run_mmrClustVar_app()
}

}
